Split isFieldWidgetEligibleNestedField into its own util file
